+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
- -   
+ -  Added ability to retrieve a sensor
  
 ### Changed
 
@@ -33,7 +33,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  -  Alert for deploying and retrieving sensors when no event selected
  -  Added ability to create a peak
  -  Added tooltips to all add button icons
- -  Added ability to retrieve a sensor
+ -  Added ability to delete a peak
  -  Added Description and Is Destroyed columns in reference datum table
  
 ### Changed
@@ -43,6 +43,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  -  Improved responsiveness of tables in details modals
  -  Autopopulate source and agency when adding file in site edit form
  -  Disabled add button when no event selected for Peaks, Sensors, HWMs and files (except site files)
+ -  Disabled peak edit when no event selected
  -  Date Recovered to Date Last Checked in reference datum detail and edit modals
  
  ### Fixed
@@ -50,6 +51,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  -  Time reset in sensor edit modal when Cancel Edits is clicked
  -  Time validation in sensor edit
  -  Inability to login after logging out
+ -  Duplicated disabled buttons when incorrect role and no event selected
 
 ## [v0.7.0](https://github.com/USGS-WiM/stnweb2/releases/tag/v0.7.0) - 2022-01-17